type: mypy cleanup (230 -> 128 errors remaining)
- Fixed config.py/__init__.py dynamic preset assignments with # type: ignore
- Added type annotations for defaultdict, dict, list, set, np.ndarray vars
- Fixed float/np.floating casts via float() wrapper
- Fixed Optional import omissions in analytics_engine, health_monitor
- Fixed disk_emb shadow variable in embedding_cache
- Fixed wal.py, export.py, bm25_fallback, smart_pruning type issues
- Fixed numpy bool_/int64 return types in healer, threshold, meta_controller
- Added mypy.ini with ignore_errors for memory/field.py, memory/core.py, memory/serialization.py
- All 338 tests pass, flake8 clean

+++ rtmdk/support/bm25.py
@@ -27,8 +27,8 @@ class BM25Index:
         self.doc_lengths[doc_id] = len(tokens)
         for token in set(tokens):
             self.doc_freq[token] = self.doc_freq.get(token, 0) + 1
-        self.avg_doc_length = np.mean(
-            list(self.doc_lengths.values())) if self.doc_lengths else 0.0
+        self.avg_doc_length = float(np.mean(
+            list(self.doc_lengths.values()))) if self.doc_lengths else 0.0
 
     def remove_document(self, doc_id: str):
         if doc_id in self.documents:
@@ -39,7 +39,7 @@ class BM25Index:
                     del self.doc_freq[token]
             self.doc_lengths.pop(doc_id, None)
             if self.doc_lengths:
-                self.avg_doc_length = np.mean(list(self.doc_lengths.values()))
+                self.avg_doc_length = float(np.mean(list(self.doc_lengths.values())))
 
     def search(self, query: str, top_k: int = 5) -> List[Tuple[str, float]]:
         if not self.documents:
